pyproject.toml
[tool.poetry]
name = "ai4se-gnusarev-1"
version = "0.1.0"
description = ""
authors = ["Sergey Gnusarev <sagnusaryov@edu.hse.ru>"]
readme = "README.md"
packages = [{include = "ai4se_gnusarev_1", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.34.0"
fastexcel = "^0.16.0"
tqdm = "^4.67.1"
xlsxwriter = "^3.2.9"
numpy = "^2.3.4"
scikit-learn = "^1.7.2"
pyyaml = "^6.0.3"
optuna = "^4.5.0"
torch = {version = "^2.9.0", source = "pytorch_cpu"}
transformers = "^4.57.1"
torchmetrics = "^1.8.2"
safetensors = "^0.6.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^7.0.1"
ruff = "^0.14.2"
ipywidgets = "^8.1.7"



[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
target-version = "py312"
